+++ cli_tools/common/utils/daisy/daisy_utils.go
@@ -63,53 +63,6 @@ func GetTranslateWorkflowPath(os *string) string {
 	return osChoices[*os]
 }
 
-// ParseWorkflow parses Daisy workflow file and returns Daisy workflow object or error in case of failure
-func ParseWorkflow(path string, varMap map[string]string, project, zone, gcsPath, oauth, dTimeout,
-	cEndpoint string, disableGCSLogs, disableCloudLogs, disableStdoutLogs bool) (*daisy.Workflow, error) {
-	w, err := daisy.NewFromFile(path)
-	if err != nil {
-		return nil, err
-	}
-Loop:
-	for k, v := range varMap {
-		for wv := range w.Vars {
-			if k == wv {
-				w.AddVar(k, v)
-				continue Loop
-			}
-		}
-		return nil, fmt.Errorf("unknown workflow Var %q passed to Workflow %q", k, w.Name)
-	}
-
-	w.Project = project
-	w.Zone = zone
-	if gcsPath != "" {
-		w.GCSPath = gcsPath
-	}
-	if oauth != "" {
-		w.OAuthPath = oauth
-	}
-	if dTimeout != "" {
-		w.DefaultTimeout = dTimeout
-	}
-
-	if cEndpoint != "" {
-		w.ComputeEndpoint = cEndpoint
-	}
-
-	if disableGCSLogs {
-		w.DisableGCSLogging()
-	}
-	if disableCloudLogs {
-		w.DisableCloudLogging()
-	}
-	if disableStdoutLogs {
-		w.DisableStdoutLogging()
-	}
-
-	return w, nil
-}
-
 // UpdateAllInstanceNoExternalIP updates all Create Instance steps in the workflow to operate
 // when no external IP access is allowed by the VPC Daisy workflow is running in.
 func UpdateAllInstanceNoExternalIP(workflow *daisy.Workflow, noExternalIP bool) {
