+++ cli_tools/common/utils/daisy/daisy_utils_test.go
@@ -15,11 +15,9 @@
 package daisyutils
 
 import (
-	"reflect"
 	"testing"
 
 	"github.com/GoogleCloudPlatform/compute-image-tools/daisy"
-	"github.com/stretchr/testify/assert"
 	"google.golang.org/api/compute/v1"
 )
 
@@ -54,38 +52,6 @@ func TestGetTranslateWorkflowPathInvalid(t *testing.T) {
 	}
 }
 
-func TestParseWorkflows(t *testing.T) {
-	path := "../../../../daisy/test_data/test.wf.json"
-	varMap := map[string]string{"key1": "var1", "key2": "var2"}
-	project := "project"
-	zone := "zone"
-	gcsPath := "gcspath"
-	oauth := "oauthpath"
-	dTimeout := "10m"
-	endpoint := "endpoint"
-	w, err := ParseWorkflow(path, varMap, project, zone, gcsPath, oauth, dTimeout, endpoint, true,
-		true, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	assertWorkflow(t, w, project, zone, gcsPath, oauth, dTimeout, endpoint, varMap)
-}
-
-func TestParseWorkflowsInvalidPath(t *testing.T) {
-	varMap := map[string]string{"key1": "var1", "key2": "var2"}
-	project := "project"
-	zone := "zone"
-	gcsPath := "gcspath"
-	oauth := "oauthpath"
-	dTimeout := "10m"
-	endpoint := "endpoint"
-	w, err := ParseWorkflow("NOT_VALID_PATH", varMap, project, zone, gcsPath, oauth, dTimeout, endpoint,
-		true, true, true)
-	assert.Nil(t, w)
-	assert.NotNil(t, err)
-}
-
 func TestUpdateWorkflowInstancesConfiguredForNoExternalIP(t *testing.T) {
 	w := createWorkflowWithCreateInstanceNetworkAccessConfig()
 	UpdateAllInstanceNoExternalIP(w, true)
@@ -111,28 +77,6 @@ func TestUpdateWorkflowInstancesNotModifiedIfNoNetworkInterfaceElement(t *testin
 
 	if (*w.Steps["ci"].CreateInstances)[0].Instance.NetworkInterfaces != nil {
 		t.Errorf("Instance NetworkInterfaces should stay nil if nil before update")
-	}
-}
-
-func assertWorkflow(t *testing.T, w *daisy.Workflow, project string, zone string, gcsPath string,
-	oauth string, dTimeout string, endpoint string, varMap map[string]string) {
-	tests := []struct {
-		want, got interface{}
-	}{
-		{w.Project, project},
-		{w.Zone, zone},
-		{w.GCSPath, gcsPath},
-		{w.OAuthPath, oauth},
-		{w.DefaultTimeout, dTimeout},
-		{w.ComputeEndpoint, endpoint},
-	}
-	for _, tt := range tests {
-		if tt.want != tt.got {
-			t.Errorf("%v != %v", tt.want, tt.got)
-		}
-	}
-	if reflect.DeepEqual(w.Vars, varMap) {
-		t.Errorf("unexpected vars, want: %v, got: %v", varMap, w.Vars)
 	}
 }
