+++ cli_tools/gce_vm_image_export/main.go
@@ -29,6 +29,7 @@ import (
 	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/path"
 	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/storage"
 	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/validation"
+	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/daisy_common"
 	"github.com/GoogleCloudPlatform/compute-image-tools/daisy"
 )
 
@@ -125,7 +126,7 @@ func buildDaisyVars() map[string]string {
 func runExport(ctx context.Context) error {
 	exportWorkflowPath := getWorkflowPath()
 	varMap := buildDaisyVars()
-	workflow, err := daisyutils.ParseWorkflow(exportWorkflowPath, varMap,
+	workflow, err := daisycommon.ParseWorkflow(exportWorkflowPath, varMap,
 		*project, *zone, *scratchBucketGcsPath, *oauth, *timeout, *ce, *gcsLogsDisabled,
 		*cloudLogsDisabled, *stdoutLogsDisabled)
 	if err != nil {
